AB#5537 refactored `user list`, mv `group [assign|unassign]` -> `group role [add|remove]`, added `group role list`
