+++ austrakka/components/group/funcs.py
@@ -1,14 +1,11 @@
-from typing import List
-
 from loguru import logger
 
 from austrakka.utils.api import api_get
 from austrakka.utils.api import api_post
 from austrakka.utils.api import api_put
-from austrakka.utils.api import api_patch
+from austrakka.utils.helpers.groups import format_group_dto_for_output
 from austrakka.utils.misc import logger_wraps
 from austrakka.utils.paths import GROUP_PATH
-from austrakka.utils.helpers.groups import format_group_dto_for_output
 
 
 @logger_wraps()
@@ -47,48 +44,6 @@ def add_group(
     return api_post(
         path=GROUP_PATH,
         data=payload)
-
-
-@logger_wraps()
-def assign_groups(
-        user_id: str,
-        group_roles: List[str]):
-
-    sub_path = "assign"
-    return change_user_group_assignment(user_id, group_roles, sub_path)
-
-
-@logger_wraps()
-def unassign_groups(
-        user_id: str,
-        group_roles: List[str]):
-
-    sub_path = "unassign"
-    return change_user_group_assignment(user_id, group_roles, sub_path)
-
-
-def change_user_group_assignment(user_id, group_roles, sub_path):
-    if len(group_roles) == 0:
-        logger.warning("Nothing to do.")
-        return None
-
-    group_role_pairs = []
-    for group_role in group_roles:
-        pairs = group_role.split(",")
-        group_role_pairs.append({
-            "groupName": pairs[0],
-            "roleName": pairs[1]
-        })
-
-    payload = {
-        "objectId": user_id,
-        "entitlements": group_role_pairs
-    }
-
-    return api_patch(
-        path=f'{GROUP_PATH}/{sub_path}',
-        data=payload,
-    )
 
 
 @logger_wraps()
